flask_app/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch==2.5.1+cpu
flask==3.1.0
gunicorn==23.0.0
minio==7.2.12
pdfplumber==0.11.4
python-docx==1.1.2
chromadb-client==1.5.2
sentence-transformers==3.3.1
requests==2.32.3
